refactor: Definition の移動
