+++ core/src/main/scala/com/iheart/playSwagger/exception/MissingBaseSpecException.scala
@@ -1,0 +1,4 @@
+package com.iheart.playSwagger.exception
+
+class MissingBaseSpecException(baseSpecFileName: String)
+    extends Exception(s"Missing a $baseSpecFileName.yml or $baseSpecFileName.json to provide base swagger spec")
