+++ core/src/test/scala/com/iheart/playSwagger/DefinitionGeneratorSpec.scala
@@ -1,8 +1,6 @@
 package com.iheart.playSwagger
-
-import com.iheart.playSwagger.Domain._
-import com.iheart.playSwagger.domain.CustomTypeMapping
 import com.iheart.playSwagger.domain.parameter.{CustomSwaggerParameter, GenSwaggerParameter}
+import com.iheart.playSwagger.domain.{CustomTypeMapping, Definition}
 import com.iheart.playSwagger.generator.{DefinitionGenerator, DomainModelQualifier, NamingConvention, PrefixDomainModelQualifier, SwaggerParameterMapper}
 import org.specs2.mutable.Specification
 import play.api.libs.json.Json
